+++ src/by/it/mialeshka/calc/ParserTest.java
@@ -3,7 +3,9 @@ package by.it.mialeshka.calc;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.runners.MethodSorters;
-import static org.junit.Assert.*;
+
+import static org.junit.Assert.assertArrayEquals;
+import static org.junit.Assert.assertEquals;
 
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
 public class ParserTest {
