+++ src/by/it/mialeshka/calc/Printer.java
@@ -1,5 +1,7 @@
 package by.it.mialeshka.calc;
 
+import by.it.mialeshka.calc.Var;
+
 class Printer {
     void print (Var var){
         if (var != null){
