+++ src/by/it/mialeshka/calc/Scalar.java
@@ -20,7 +20,7 @@ class Scalar extends Var {
     }
 
     @Override
-    public Var add(Var other) throws CalcException{
+    public Var add(Var other) throws CalcException {
         if(other instanceof Scalar)
         {
             double sum = this.var+((Scalar) other).var;
@@ -30,7 +30,7 @@ class Scalar extends Var {
     }
 
     @Override
-    public Var sub(Var other) throws CalcException{
+    public Var sub(Var other) throws CalcException {
         if(other instanceof Scalar)
         {
             double sub = this.var-((Scalar) other).var;
@@ -40,7 +40,7 @@ class Scalar extends Var {
     }
 
     @Override
-    public Var mul(Var other) throws CalcException{
+    public Var mul(Var other) throws CalcException {
         if(other instanceof Scalar)
         {
             double mul = this.var*((Scalar) other).var;
@@ -55,11 +55,11 @@ class Scalar extends Var {
 
 
     @Override
-    public Var div(Var other) throws CalcException{
+    public Var div(Var other) throws CalcException {
         if(other instanceof Scalar)
         {
             if(((Scalar) other).var == 0)
-                throw new CalcException("Деление на ноль");
+                throw new CalcException(ErrMesages.res.get(ErrMesages.divZero));
             double div = this.var/((Scalar) other).var;
             return new Scalar(div);
         }
