+++ src/by/it/mialeshka/calc/VarTest.java
@@ -2,7 +2,7 @@ package by.it.mialeshka.calc;
 
 import org.junit.Test;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 public class VarTest {
