+++ src/by/it/mialeshka/calc/Vector.java
@@ -1,12 +1,11 @@
 package by.it.mialeshka.calc;
 
+
 import java.util.Arrays;
 import java.util.regex.Pattern;
 
 class Vector extends Var {
     double[] var;
-
-
 
     public Vector(double[] var) {
         this.var = var;
@@ -42,7 +41,7 @@ class Vector extends Var {
     }
 
     @Override
-    public Var add(Var other) throws CalcException{
+    public Var add(Var other) throws CalcException {
         if (other instanceof Scalar){
             double[] res = Arrays.copyOf(var, var.length);
             for (int i = 0; i < res.length; i++) {
@@ -52,7 +51,7 @@ class Vector extends Var {
         }
         else if (other instanceof Vector){
             if(this.var.length != ((Vector) other).var.length)
-                throw new CalcException("Длинна векторов не равна");
+                throw new CalcException(ErrMesages.res.get(ErrMesages.lengthVectorErr));
             double[] res = Arrays.copyOf(var, var.length);
             for (int i = 0; i < res.length; i++) {
                 res[i] = res[i] + ((Vector) other).var[i];
@@ -64,7 +63,7 @@ class Vector extends Var {
     }
 
     @Override
-    public Var sub(Var other) throws CalcException{
+    public Var sub(Var other) throws CalcException {
         if (other instanceof Scalar){
             double[] res = Arrays.copyOf(var, var.length);
             for (int i = 0; i < res.length; i++) {
@@ -74,7 +73,7 @@ class Vector extends Var {
         }
         else if (other instanceof Vector){
             if(this.var.length != ((Vector) other).var.length)
-                throw new CalcException("Длинна векторов не равна");
+                throw new CalcException(ErrMesages.res.get(ErrMesages.lengthVectorErr));
             double[] res = Arrays.copyOf(var, var.length);
             for (int i = 0; i < res.length; i++) {
                 res[i] = res[i] - ((Vector) other).var[i];
@@ -86,7 +85,7 @@ class Vector extends Var {
     }
 
     @Override
-    public Var mul(Var other) throws CalcException{
+    public Var mul(Var other) throws CalcException {
         if (other instanceof Scalar){
             double[] res = Arrays.copyOf(var, var.length);
             for (int i = 0; i < res.length; i++) {
@@ -97,7 +96,7 @@ class Vector extends Var {
         else if (other instanceof Vector){
             double res = 0.0;
             if(this.var.length != ((Vector) other).var.length)
-                throw new CalcException("Длинна векторов не равна");
+                throw new CalcException(ErrMesages.res.get(ErrMesages.lengthVectorErr));
             for (int i = 0; i < this.var.length; i++) {
                 res = res + this.var[i]*((Vector) other).var[i];
             }
@@ -107,10 +106,10 @@ class Vector extends Var {
     }
 
     @Override
-    public Var div(Var other) throws CalcException{
+    public Var div(Var other) throws CalcException {
         if (other instanceof Scalar){
             if(((Scalar) other).var == 0)
-                throw new CalcException("Деление на ноль");
+                throw new CalcException(ErrMesages.res.get(ErrMesages.divZero));
             double[] res = Arrays.copyOf(var, var.length);
             for (int i = 0; i < res.length; i++) {
                 res[i] = res[i] / ((Scalar) other).getVar();
