+++ src/by/it/mialeshka/calc/res/language.properties
