+++ src/by/it/mialeshka/calc/res/language_be_BY.properties
