+++ src/by/it/mialeshka/calc/res/language_en_US.properties
