+++ src/by/it/mialeshka/calc/txt/language.txt
