+++ src/by/it/mialeshka/calc/vars.txt
