--- src/by/it/mialeshka/jd02_05/calc/CalcException.java
@@ -1,23 +1,0 @@
-package by.it.mialeshka.jd02_05.calc;
-
-public class CalcException extends Exception{
-    public CalcException() {
-        super();
-    }
-
-    public CalcException(String message) {
-        super("ERROR: " + message);
-    }
-
-    public CalcException(String message, Throwable cause) {
-        super("ERROR: " + message, cause);
-    }
-
-    public CalcException(Throwable cause) {
-        super(cause);
-    }
-
-    protected CalcException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
-        super("ERROR: " + message, cause, enableSuppression, writableStackTrace);
-    }
-}
