--- src/by/it/mialeshka/jd02_05/calc/Operation.java
@@ -1,8 +1,0 @@
-package by.it.mialeshka.jd02_05.calc;
-
-public interface Operation {
-    Var add(Var other) throws CalcException;
-    Var sub(Var other) throws CalcException;
-    Var mul(Var other) throws CalcException;
-    Var div(Var other) throws CalcException;
-}
