--- src/by/it/mialeshka/jd02_05/calc/Patterns.java
@@ -1,9 +1,0 @@
-package by.it.mialeshka.jd02_05.calc;
-
-interface Patterns {
-    String OPERATION = "(?<=[^-+*=/{,])[-+*/=]";
-    String SCALAR = "-?[0-9]+\\.?[0-9]*";
-    String VECTOR = "\\{"+SCALAR+"(,\\s*"+SCALAR+")*}";
-    String MATRIX = "\\{"+VECTOR+"(,\\s*"+VECTOR+")*}";
-    String GROUPOPERATION  = "[(]{1}[-+*/0-9a-zA-Z.,{}]+[)]{1}";
-}
