--- src/by/it/mialeshka/jd02_05/calc/Printer.java
@@ -1,9 +1,0 @@
-package by.it.mialeshka.jd02_05.calc;
-
-class Printer {
-    void print (Var var){
-        if (var != null){
-           System.out.println(var);
-        }
-    }
-}
